When checking in, transfer excess balance to the owner
This is useful because, when checking in, the owner cannot precisely calculate how much they need to transfer to the contract in order to cover all Keepers' fees. This is because, at the moment of transaction generation, owner doesn't know which block this transaction will be included into, which means that the presice value of `now` is not known. And we're crediting Keepers in proportion with the amount of time passed since last owner check-in. This means that, in order to guarantee successful check-in, owner needs to send a bit more than the value they calculated locally. So the guarantee that any excess funds will be sent back to the owner's address is a nice thing to have.